test: use qunit syntax
Personal preference, more closely resembles unit test style
