+++ vscode-extension/test/e2e/defaults-scss/completion-sassdoc.test.js
@@ -1,0 +1,77 @@
+const { getDocUri, showFile, position, sleepCI } = require("./util");
+const { testCompletion } = require("./completion-helper");
+
+const docUri = getDocUri("completion/sassdoc.scss");
+
+before(async () => {
+	await showFile(docUri);
+});
+
+test("Offers completions for SassDoc block on mixin without parameters or @content", async () => {
+	const expectedCompletions = [
+		{
+			label: "SassDoc Block",
+			insertText: '" ${0}\\n/// @output ${1}"',
+		},
+	];
+
+	await testCompletion(docUri, position(3, 4), expectedCompletions);
+});
+
+test("Offers completions for SassDoc block on mixin with @content", async () => {
+	const expectedCompletions = [
+		{
+			label: "SassDoc Block",
+			insertText: '" ${0}\\n/// @content ${1}\\n/// @output ${2}"',
+		},
+	];
+
+	await testCompletion(docUri, position(8, 4), expectedCompletions);
+});
+
+test("Offers completions for SassDoc block on mixin with parameters", async () => {
+	const expectedCompletions = [
+		{
+			label: "SassDoc Block",
+			insertText:
+				'" ${0}\\n/// @param {${1:Number}} \\\\$a [1px] ${2:-}\\n/// @param {${3:Number}} \\\\$b [2px] ${4:-}\\n/// @output ${5}"',
+		},
+	];
+
+	await testCompletion(docUri, position(13, 4), expectedCompletions);
+});
+
+test("Offers completions for SassDoc block on mixin with parameters and @content", async () => {
+	const expectedCompletions = [
+		{
+			label: "SassDoc Block",
+			insertText:
+				'" ${0}\\n/// @param {${1:type}} \\\\$a ${2:-}\\n/// @param {${3:type}} \\\\$b ${4:-}\\n/// @content ${5}\\n/// @output ${6}"',
+		},
+	];
+
+	await testCompletion(docUri, position(18, 4), expectedCompletions);
+});
+
+test("Offers completions for SassDoc block on parameterless function", async () => {
+	const expectedCompletions = [
+		{
+			label: "SassDoc Block",
+			insertText: '" ${0}\\n/// @return {${1:type}} ${2:-}"',
+		},
+	];
+
+	await testCompletion(docUri, position(25, 4), expectedCompletions);
+});
+
+test("Offers completions for SassDoc block on parameterfull function", async () => {
+	const expectedCompletions = [
+		{
+			label: "SassDoc Block",
+			insertText:
+				'" ${0}\\n/// @param {${1:Number}} \\\\$a [1px] ${2:-}\\n/// @param {${3:Number}} \\\\$b [2px] ${4:-}\\n/// @return {${5:type}} ${6:-}"',
+		},
+	];
+
+	await testCompletion(docUri, position(30, 4), expectedCompletions);
+});
