+++ vscode-extension/test/e2e/suggest-from-use-only/suggest-from-use-only.test.js
@@ -1,45 +1,51 @@
 const { showFile, position, sleepCI } = require("../util");
 const { testCompletion, getDocPath } = require("./helper");
 
-describe("Suggest from use only", function () {
-	const styles = getDocPath("styles.scss");
+const styles = getDocPath("styles.scss");
 
-	before(async () => {
-		await showFile(styles);
-		await sleepCI();
-	});
+before(async () => {
+	await showFile(styles);
+	await sleepCI();
+});
 
-	it("setting gets applied", async () => {
-		await testCompletion(styles, position(3, 6), [
-			{
-				label: "$a",
-			},
-		]);
+test("setting gets applied", async () => {
+	await testCompletion(styles, position(3, 6), [
+		{
+			label: "$a",
+		},
+	]);
 
-		// Does not suggest from b or c
-		await testCompletion(
-			styles,
-			position(3, 6),
-			[
-				{
-					label: "$b",
-				},
-			],
+	// Does not suggest from b or c
+	await testCompletion(
+		styles,
+		position(3, 6),
+		[
 			{
-				expectNoMatch: true,
+				label: "$b",
 			},
-		);
-		await testCompletion(
-			styles,
-			position(3, 6),
-			[
-				{
-					label: "$c",
-				},
-			],
+		],
+		{
+			expectNoMatch: true,
+		},
+	);
+	await testCompletion(
+		styles,
+		position(3, 6),
+		[
 			{
-				expectNoMatch: true,
+				label: "$c",
 			},
-		);
+		],
+		{
+			expectNoMatch: true,
+		},
+	);
+});
+
+test("offers no hidden items in namespace completions", async () => {
+	let expectedCompletions = ["$secret"];
+
+	await testCompletion(styles, position(23, 13), expectedCompletions, {
+		expectNoMatch: true,
 	});
 });
