+++ vscode-extension/test/e2e/suggest-from-use-only/workspace/_a.scss
@@ -1,1 +1,3 @@
+@forward "d" hide $secret;
+
 $a: 1;
