+++ vscode-extension/test/e2e/suggest-from-use-only/workspace/_d.scss
@@ -1,0 +1,1 @@
+$secret: 1;
